Refactor codegen
