+++ src/typedefs.jl
@@ -1,0 +1,16 @@
+mutable struct VLSpec{T}
+    params::Union{Dict, Vector}
+end
+
+vltype(::VLSpec{T}) where {T} = T
+
+### list of VegaLite property names that need a new denomination in Julia
+const sp2jl = Dict{Symbol,Symbol}(:type => :typ)
+const jl2sp = Dict( (v,k) for (k,v) in sp2jl)
+
+const arrayprops = Symbol[:layer, :transform, :hconcat, :vconcat]
+
+### conversion between property name in VegaLite and julia function name
+jlfunc(vln::String) = jlfunc(Symbol(vln))
+jlfunc(vln::Symbol) = Symbol("vl" * string(vln))
+vlname(fn::Symbol)  = replace(string(fn), r"^vl", "")
